+++ bioframe/__init__.py
@@ -1,5 +1,8 @@
 from ._version import __version__
 from .io import *
-from .region import *
+from .core.stringops import *
+from .core.specs import *
+from .core.construction import *
 from .ops import *
 from .genomeops import *
+from .core.checks import *
