+++ bioframe/io/fileops.py
@@ -18,7 +18,7 @@ except ImportError:
     pyBigWig = None
 
 from ..util import run
-from ..region import parse_region
+from ..core.stringops import parse_region
 from ..arrops import argnatsort
 from .schemas import SCHEMAS, BAM_FIELDS, GAP_FIELDS, UCSC_MRNA_FIELDS
