+++ bioframe/io/resources.py
@@ -13,7 +13,7 @@ import glob
 import pkg_resources
 
 from .schemas import SCHEMAS
-from .formats import (
+from .fileops import (
     read_table,
     read_chromsizes,
     read_gapfile,
