+++ docs/conf.py
@@ -12,15 +12,17 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../'))
+
+sys.path.insert(0, os.path.abspath("../"))
 
 import mock
+
 MOCK_MODULES = [
-    'numpy',
-    'pandas',
-    'pandas.algos',
-    'pandas.api',
-    'pandas.api.types',
+    "numpy",
+    "pandas",
+    "pandas.algos",
+    "pandas.api",
+    "pandas.api.types",
     # 'dask',
     # 'dask.base',
     # 'dask.array',
@@ -33,15 +35,16 @@ for mod_name in MOCK_MODULES:
 
 # -- Project information -----------------------------------------------------
 
-project = 'bioframe'
-copyright = '2020, OmicRelief'
-author = 'OmicRelief'
+project = "bioframe"
+copyright = "2020, OmicRelief"
+author = "OmicRelief"
 
 # The full version, including alpha/beta/rc tags
 def _read(*parts, **kwargs):
     import os, io
+
     filepath = os.path.join(os.path.dirname(__file__), *parts)
-    encoding = kwargs.pop('encoding', 'utf-8')
+    encoding = kwargs.pop("encoding", "utf-8")
     with io.open(filepath, encoding=encoding) as fh:
         text = fh.read()
     return text
@@ -49,11 +52,14 @@ def _read(*parts, **kwargs):
 
 def get_version():
     import re
+
     version = re.search(
         r'^__version__\s*=\s*[\'"]([^\'"]*)[\'"]',
-        _read('..','bioframe', '_version.py'),
-        re.MULTILINE).group(1)
+        _read("..", "bioframe", "_version.py"),
+        re.MULTILINE,
+    ).group(1)
     return version
+
 
 version = get_version()
 # The full version, including alpha/beta/rc tags.
@@ -65,24 +71,24 @@ release = version
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
+    "sphinx.ext.autodoc",
     # 'sphinx.ext.doctest',
-    # 'sphinx.ext.todo',                                                                                                      
+    # 'sphinx.ext.todo',
     #'sphinx.ext.coverage',
     # 'sphinx.ext.mathjax',
     # 'sphinx.ext.ifconfig',
-    'autodocsumm',
-    'sphinx.ext.viewcode',
-    'sphinx.ext.autosummary',                                                                                              
-    'sphinx.ext.napoleon',  # 'numpydoc'
+    "autodocsumm",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.napoleon",  # 'numpydoc'
 ]
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -90,13 +96,13 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
-master_doc = 'index'
+master_doc = "index"
 
-autosummary_generate = True 
+autosummary_generate = True
