+++ setup.cfg
@@ -1,2 +1,2 @@
 [bdist_wheel]
-universal=1
+universal = 1
